001_accessing_claude_with_the_api/004_system_prompts/requirements.txt
anthropic
python-dotenv
streamlit